Removed the core module. Moved everything to worker module
