+++ spatialos-sdk/src/worker/component.rs
