+++ spatialos-sdk/src/worker/connection.rs
@@ -1,11 +1,11 @@
 use std::ffi::{CStr, CString};
 use std::ptr;
-use worker::core::commands::*;
-use worker::core::component::ComponentUpdate;
-use worker::core::metrics::Metrics;
-use worker::core::op::{DisconnectOp, OpList, WorkerOp};
-use worker::core::parameters::{CommandParameters, ConnectionParameters};
-use worker::core::{EntityId, InterestOverride, LogLevel, RequestId};
+use worker::commands::*;
+use worker::component::ComponentUpdate;
+use worker::metrics::Metrics;
+use worker::op::{DisconnectOp, OpList, WorkerOp};
+use worker::parameters::{CommandParameters, ConnectionParameters};
+use worker::{EntityId, InterestOverride, LogLevel, RequestId};
 use worker::internal::bindings::*;
 
 /// Connection trait to allow for mocking the connection.
