+++ spatialos-sdk/src/worker/entity_snapshot.rs
