+++ spatialos-sdk/src/worker/vtable.rs
